src/implementation_timeline_4.py
# Implementation Timeline
import streamlit as st
import pandas as pd
import altair as alt
import plotly.express as px

def implementation_timeline():
    st.markdown('<p class="big-font">Implementation Timeline</p>', unsafe_allow_html=True)
    st.write("Efficient deployment in 4 months")

    # Timeline chart
    timeline_data = [
        dict(Task="Test Integration", Start='2024-08-07', Finish='2024-08-28', Resource='Stage 1'),
        dict(Task="Database Connection", Start='2024-08-21', Finish='2024-09-11', Resource='Stage 1'),
        dict(Task="API Integration", Start='2024-09-04', Finish='2024-10-16', Resource='Stage 2'),
        dict(Task="Testing and QA", Start='2024-10-02', Finish='2024-11-13', Resource='Stage 3'),
        dict(Task="Final Deployment", Start='2024-11-06', Finish='2024-11-27', Resource='Stage 4'),
    ]

    fig = px.timeline(timeline_data, x_start="Start", x_end="Finish", y="Task", color="Resource")
    fig.update_layout(title="Integration Timeline", xaxis_title="Date", yaxis_title="Task")
    st.plotly_chart(fig)

    # Integration components
    st.subheader("Key Integration Components")
    col1, col2 = st.columns(2)

    with col1:
        st.markdown("""
        ### 1. Website Integration
        - Embed chatbot widget
        - Customize UI to match brand
        - Implement responsive design
        """)

        st.markdown("""
        ### 2. Database Connection
        - Secure property database link
        - Real-time data synchronization
        - Data encryption and protection
        """)

    with col2:
        st.markdown("""
        ### 3. API Integration
        - MLS API connection
        - Third-party service APIs
        - Data consistency protocols
        """)

        st.markdown("""
        ### 4. Testing and QA
        - Comprehensive integration tests
        - Performance optimization
        - User acceptance testing
        """)
